[scripts][run-zircon] switch arm64 qemu to start using zircon.bin
This is in prep for removing the MEMBASE variable which is currently
needed to link the kernel at the proper physical address for qemu. If we
load zircon via the .bin file it doesn't use the ELF headers so we dont
need to link the kernel to any particular address.

Change-Id: I38b99c2e65d757df762602350c2e26b3fc5c3d62